oh jaaaaaaaaaaaaaaaa, VECTOR!!!

+++ Vireo/src/io/memory.h
@@ -16,7 +16,7 @@ void GRUB_GetMemInfo(multiboot_info_t* mbh);
 
 void *kmemset(void *ptr, uint32_t val, size_t size);
 
-void MemCopy(uint32_t *src, uint32_t *dest, size_t size);
+void MemCopy(uint32_t *src, uint32_t *dest, uint32_t size);
 char *strcopy(char* dest, char* src);
 
 //#define KMEMSET_KOTHER 0xB17D05 //BirdOS
